requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
executing==2.2.0
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
mysqlclient==2.2.7
packaging==25.0
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-decouple==3.8
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.9.0
